+++ log\.6c1a1f7b9023a40cab0375214d702bb453b1bc33-audit.json
@@ -1,0 +1,35 @@
+{
+    "keep": {
+        "days": true,
+        "amount": 14
+    },
+    "auditLog": "log\\.6c1a1f7b9023a40cab0375214d702bb453b1bc33-audit.json",
+    "files": [
+        {
+            "date": 1658884242573,
+            "name": "log\\2022-07-27-smart-push.log",
+            "hash": "f240af5006398fe86ebeb032c4ed798ad4b646b604498d8313f3fdde622fbacc"
+        },
+        {
+            "date": 1659083503623,
+            "name": "log\\2022-07-29-smart-push.log",
+            "hash": "b43366692d3fb06d360a2c0e9c55c697c31469cd245fa64ccaf0213162b8fd7c"
+        },
+        {
+            "date": 1659157091090,
+            "name": "log\\2022-07-30-smart-push.log",
+            "hash": "2510ddeb4aac2e7b9ec4d13e427f9ed36b424df4d1f95aa662e132088ad654bb"
+        },
+        {
+            "date": 1659278376037,
+            "name": "log\\2022-07-31-smart-push.log",
+            "hash": "ec46c59ea00d75a3b4906a6e9f378b962603c3fb7891c31606d5e0c1e7b60623"
+        },
+        {
+            "date": 1659454459015,
+            "name": "log/2022-08-03-smart-push.log",
+            "hash": "3bd6ee77556ed8b520e8f7c13fb3d4025f28dbf11ce5c23f8320e034594b1f09"
+        }
+    ],
+    "hashType": "sha256"
+}
