+++ src/app/Group/groupDao.js
@@ -1,0 +1,12 @@
+const selectUserPosts = async (connection) => {
+    const selectTestUserQuery = `
+          SELECT *
+          FROM User
+          ;
+        `;
+    const [testResult] = await connection.query(selectTestUserQuery);
+  
+    return testResult;
+  };
+  
+  module.exports = { selectUserPosts };
