+++ src/app/Group/groupRoute.js
@@ -1,0 +1,8 @@
+import express from "express";
+import { getDatabaseTest } from "./groupController";
+
+const groupRouter = express.Router();
+
+groupRouter.get("/db", getDatabaseTest);
+export default groupRouter;
+
