+++ src/app/Member/memberController.js
@@ -1,0 +1,12 @@
+const baseResponse = require("../../../config/baseResponseStatus");
+const { response, errResponse } = require("../../../config/response");
+const memberProvider = require("./memberProvider");
+
+
+
+
+export const getDatabaseTest = async (req, res) => {
+    const testUserResult = await memberProvider.retrieveUserList();
+    return res.send(testUserResult);
+  };
+
