+++ src/app/Member/memberProvider.js
@@ -1,0 +1,21 @@
+const { pool } = require("../../../config/database");
+const baseResponseStatus = require("../../../config/baseResponseStatus");
+const { response, errResponse } = require("../../../config/response");
+const { logger } = require("../../../config/winston");
+
+const memberDao = require("./memberDao");
+
+exports.retrieveUserList = async function () {
+    const connection = await pool.getConnection(async (conn) => conn);
+    const handleError = (error) => logger.error(`❌DB Error: ${error.message}`);
+  
+    try {
+      const testResult = await memberDao.selectUserPosts(connection);
+      connection.release();
+      return response(baseResponseStatus.SUCCESS, testResult);
+    } catch (error) {
+      handleError(error);
+      connection.release();
+      return errResponse(baseResponseStatus.FAILURE);
+    }
+  };
