+++ src/app/Member/memberRoute.js
@@ -1,0 +1,8 @@
+import express from "express";
+import { getDatabaseTest } from "./memberController";
+
+const memberRouter = express.Router();
+
+memberRouter.get("/db", getDatabaseTest);
+
+export default memberRouter;
