--- src/app/TestInit/testDao.js
@@ -1,12 +1,0 @@
-const selectUserPosts = async (connection) => {
-  const selectTestUserQuery = `
-        SELECT *
-        FROM TestUser
-        ;
-      `;
-  const [testResult] = await connection.query(selectTestUserQuery);
-
-  return testResult;
-};
-
-module.exports = { selectUserPosts };
